pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pillow>=11.0.0,<12.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "python-multipart>=0.0.18,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]
ocr-deepseek = [
    "torch>=2.4.0,<3.0.0",
    "transformers>=4.46.0,<5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
